ISSUE-27 Merge module_util class files
