+++ plugins/module_utils/service_now.py
@@ -26,9 +26,25 @@ HAS_REQUESTS = False
 REQUESTS_IMP_ERR = None
 try:
     import requests
+    from requests.auth import AuthBase
     HAS_REQUESTS = True
 except ImportError:
     REQUESTS_IMP_ERR = traceback.format_exc()
+
+
+class HTTPBearerAuth(AuthBase):
+    """A :class:`requests.auth` bearer token authentication method
+    per https://2.python-requests.org/en/master/user/authentication/#new-forms-of-authentication
+
+    :param token: Bearer token to be used instead of user/pass or session
+    """
+
+    def __init__(self, token):
+        self.token = token
+
+    def __call__(self, r):
+        r.headers['Authorization'] = "Bearer {0}".format(str(self.token))
+        return r
 
 
 class ServiceNowModule(AnsibleModule):
